create .h .c
