+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Jun 03 15:10 Egypt Standard Time
+Start testing: Jun 03 15:24 Egypt Standard Time
 ----------------------------------------------------------
-End testing: Jun 03 15:10 Egypt Standard Time
+End testing: Jun 03 15:24 Egypt Standard Time
